backend/pyproject.toml
[build-system]
requires = ["uv_build==0.7.19"]
build-backend = "uv_build"

[project]
name = "h_rag"
version = "0.1"
readme = "README.md"
requires-python = "==3.13.12"
dependencies = [
    "boto3>=1.42.59",
    "chromadb>=1.5.0",
    "einops>=0.8.2",
    "en-core-web-sm",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.63.0",
    "loguru>=0.7.3",
    "ollama>=0.6.1",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.3",
    "spacy>=3.8.11",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "pytest==9.0.2",
    "pytest-mock>=3.15.1",
    "ruff==0.15.6",
    "ty>=0.0.24",
]

[tool.fastapi]
entrypoint = "src.h_rag.api.main:app"

[tool.pytest]
addopts = ["--import-mode=importlib"]
testpaths = []
log_cli = false

[tool.ruff]
include = ["src/**.py"]
exclude = ["src/**/__init__.py"]
line-length = 100

[tool.ruff.lint]
select = ["D", "F401", "I"]
ignore = ["D107"]


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
